+++ backend/src/main/java/com/gradesave/backend/dto/grade/CalculateSubjectGradeDto.java
@@ -1,6 +1,4 @@
 package com.gradesave.backend.dto.grade;
 
-import java.math.BigDecimal;
-
-public record CalculateSubjectGradeDto(BigDecimal grade, Double weight) {
+public record CalculateSubjectGradeDto(Double grade, Double weight) {
 }
